Adding different viral factor to the individuals, hospital data and possibility of not having all parameters in CSV epi
Individuals have now the possibility to have a viral factor, which will determine their beta and viral release (i.e. allowing superspreaders, for instance). The enhancement is set to false by default. Also changed the order of creation of the matrix of parameters to allow epidemiological CSV file not having all the parameters. In addition, I've added the data for the hospitalisation and ICU capacity, according to Vinh Phuc (hence why it should be added as a parameter LATER)